refactor: splitted the inspector widgets into sub widgets for maintainability and readability
